Restore ability to run the non-dev server with local data
`npm run server localData` (which is `node server.dist.js localData`)
was not working because during the webpack build process the __dirname
reference came to refer to the repo root instead of src/server/.  This
meant the server looked for a data dir two levels up.  In dev mode, the
webpack hot-reloading managed to keep __dirname correct.

This change adjusts the webpack configuration to preserve the __dirname
of the *original* file (src/server/globals.js) instead of using the
dirname of the *packed output* file (server.dist.js).

That config change alone fixes the non-dev server with local data, but
breaks the dev server with local data since the LOCAL_DATA_PATH became
relative instead of absolute.  This necessitated resolving the path to
an absolute one.

__filename is also treated the same way for good measure, although it is
not referred to in our own code.

For more details on the webpack config in play here, refer to:

   https://webpack.js.org/configuration/node/#node-__dirname