immediately remove CIs onBranchLeave to prevent d3 transition bugs
Problem: D3 selection, with transitions, seems to be lazily evaluated such at it would grab the newly displayed CI as well and remove it. Evaluating .selection() prior to setTimeout didn't fix this. This wasn't a problem before as we just added CIs in different DOM groups.

Solution: instantaneously remove the CIs on branch leave.