Refactor color by genotype state parsing
Consolidates test/encoding/decoding logic into three main functions
which are used throughout the codebase instead of hardcoded assumptions
about the serialization format ("gt-<gene>_<pos>") in many locations.
This resolves existing and prevents future divergence in the serialized
format, while allowing it to change in the future with minimal or no
change outside of getGenotype.js.

Multiple positions are supported, but multiple genes are not at this
time.  I've left that for future work since there is not a UI for
multiple genes.  This refactoring will make it easier if we do.
Terminology (prot vs. gene) was normalized for consistency when trivial.

As a result of improved parsing, gene names with embedded underscores
and hyphens are now supported.  Resolves #669.