Refactor the magic annotation key "nuc" into a global constant
It's good practice to avoid hardcoding such magic values into a myriad
of places and makes maintenance easier.  The name also lends more
context to its usage.

Using a global constant also means there is only a single place to
change the value, at least in auspice, if we need to.  It's not unlikely
we'll need to change it, because there are plenty of legitimate genes
named "nuc", e.g. <https://www.ncbi.nlm.nih.gov/gene/?term=nuc>.

Note that this doesn't refactor _all_ references to the string "nuc", as
the same value is also used for a different concept, the "mutation type"
or "mutType", which is valued "nuc" or "aa" (or sometimes false).  It's
important that magic value constants aren't used merely for their value
alone, but for the conceptual thing they represent so as not to conflate
changes in the future.