controls/color-by: Reset and replace state entirely when switching color by field
It is easier to reason about and keep track of whole state replacement
than the combined effects of individual field updates.

This is especially true for logically linked state like geneSelected and
positionSelected, which easily becomes out of sync with colorByGenotype.
For example, geneSelected and positionSelected were not reset when
switching from a genotype coloring to any other color by field.  A few
individual field updates remain where it doesn't make any more sense to
think about them in terms of replacement.