Upgrade Babel
Babel upgraded to version 7.
This allows us to use babel.config.js to have more
control over the transpiling behavior.
This allows extension JS files living outside of the
auspice directory to be processed by babel.

There is still a bug related to react-hot-loader
when using JS files outside of the auspice dir.