Enable source maps for debugging production builds
Though the development webpack config includes source maps, they were
disabled in production. When debugging a live build (or a local build
where you don't want hot reloading), it's very useful to have a source
map for the browser to translate between the packed dist/bundle.js and
the original code.

This matches the recommendations of webpack's documentation:

    https://webpack.js.org/guides/production/#source-mapping