implement styled components (CSS-in-JS)
Begin using styled components in auspice to define CSS-in-JS
and to utilise a theme which is inherited by multiple components.
Future commits will allow this theme to be provided via the
extension architecture.

Styled components was chosen because it is popular, well supported
and uses pure CSS (rather than a similar-but-different syntax).

Future commits will gradually remove our usage of global defined
styles and seperate CSS files and move towards consistent use
of styled components with themes.

in the sidebar to define styles and
supply them