Client obtains available datasets via an explicit fetch
Previously the available datasets were included from the server
in the response to a "get dataset" request.
This forced the server-provided dataset JSON to deviate from the
published JSON spec.
As auspice becomes used in different websites it's important that the
JSON delivered to the client mirrors that delivered to the server.