actions/colors: Dispatch mutType change before colorBy change
…instead of bundling the former with the latter.

Changing the mutType must happen _before_ updating colors because the
entropy bars need to be recomputed for the new mutType before applying
the new genotype colorBy.  Otherwise, the entropy component tries to
apply the new genotype colorBy to bars of the wrong mutType, which in
turn causes all sorts of errors ("entropy out of sync" and selected
positions not matching the data bars).

The state dependencies are a bit tangled here, but de-tangling them is a
larger project for another time.