reinstate full server dataset functionality
During development of the "extend" branch the second tree & tip-frequency fetching had been removed from the server. Full functionality is now restored. (Note that this code is completely separate from that in the nextstrain.org server. We could one day expose this and import it in the nextstrain.org repo.)