Regenerate docs after link fixes
The footer was affected, hence all the pages were updated.

As best I can tell without digging into the Docusaurus code, the
shifting of last updated dates in each footer seems to be because
Docusaurus uses the local timezone of whatever system it is running on
instead of a fixed timezone.  Thus builds made by James in NZ will be a
day ahead of builds done by those of us in Seattle.  Oy vey!  There's no
clear timezone configuration to set in siteConfig.js.