Fix URL query corruption when changing datasets
As described in #691, the query part of the URL was corrupted when
switching between datasets, e.g. coloring an HA tree by antigenic
advance and switching the segment to NA.

I bisected the bug to "Load datasets from a separate component"
(e026df9) and traced the problem to the query part being passed to the
PAGE_CHANGE action action as a string instead of a parsed object.
Notably, the bug existed in the previous version of the code as well,
but was harder to trigger.  Refactoring in the commit above made the bug
much easier to trigger and much more noticeable.

Resolves #691.