check mutType (aa/nuc) against mutations in nodes
Previously we assumed that  aa mutations were present and defaulted to this (currently unable to be changed by the URL query). This commit checks the default mutType ("aa") against the nodes, falling back to "nuc" if available, or "null" if neither are present. 
Additionally, we now collect available node attrs across all nodes, not just the root node, which results in an increased and more accurate set of attrs. 
Performance penalty on my laptop: 1-3ms.