Move author information (on nodes) to v2 spec
There are a lot of changes in this commit, simply as a result of moving `node.attr.authors` -> `node.authors` (v1 -> v2 JSON).
The reason is twofold: 
(a) the author key is no longer the author string to display, and it needs to be looked up via `authorInfo`
(b) Code previously relied on things being set on `node.attr[name]` and this commit begins the process of looking for these keys on both the node (`node[name]`) and in the (not yet present) traits dict (`node.traits[name]`).