Allow servers to customise auspice URL
A (completely optional) JSON field `auspice_url_should_be` will instruct the client to change the URL. This is not part of the JSON schema, but can be used by servers to mimic redirects. Note that I first implemented this using a custom HTTP header but there are issues with it being cached, so this option is both simpler and cleaner.

This commit also fixes some issues where adding a 2nd tree would sometimes not show up in the URL.