Map: refactor maybeUpdateDemesAndTransmissions
Previously, in the function `maybeUpdateDemesAndTransmissions()` (part
of the Map component), there was a large, indented code block inside of
a conditional checking for truthiness in `colorOrVisibilityChange &&
haveData`. Because changes only occurred if this conditional was true,
refactor it so that the conditional is now a guard statement. If the
conditional evaluates to false, exit the function. Now the bulk of the
code is not indented.