Map: fix bug re: geo resolution changes. Closes #722
Previously, when the geo resolution would change,
`drawDemesAndTransmissions` would not
necessarily be called, often causing data on the map
to not load at all between queries.

James also did some work here to correct the deme
and transmission data that were being loaded on
each narrative page change.

This commit further complicates the map code, but is
necessary to fix this bug. It is a prime candidate
for refactoring in #735

Co-authored-by: James Hadfield <jh22@sanger.ac.uk>
Co-authored-by: Jover Lee <joverlee521@gmail.com>