Do not reset map position when geo resolution / pie-circle changes
Previously, when you changed the geoResolution or changed between pie charts & colour-blended circles the removal of D3 elements (`maybeRemoveAllDemesAndTransmissions`) would also set `state.boundsSet` to false, causing the re-rendering of elements to also reset the map. This is an improvement as the map no longer unexpectedly moves when you change these things.