refactor `getDemeColors` into `getColorsForAllDemes`
This both improves the code quality and flow, and reports the total observed colour counts in addition to the currently-visible colour counts. This information will be necessary in order to create arcs in a pie chart for all colours at each location, whether there are any visible or not. This commit also restores the update functionality (i.e. vis / color by changes) for non-pie charts.