Unfurl clades during unrooted clade zoom
This commit alters unrooted tree layout to give more radial territory to tips that are "inView" compared to tips that are out of view. This causes clade zoom to unfurl clades when zooming. Besides altering unrootedLayout, this commit adds an "updateLayout" flag that causes layouts to be recalculated on clade zoom to account for this.