Update server conversion function to GFF format
The annotations within the metadata JSON is converted
to GFF format, with the main changes being the start
position is now one-based instead of zero-based and 
the forward/reverse strands are represented with "+" and "-".

The GFF format also includes a `seqid` and `type` key, but
this information is not available within the v1 metadata JSON, 
so they are not included in this conversion process.