Update data-getters in `loadData.js`
Currently, when the auspice client sends a partial URL to the `/charon`
API, the server injects a magic key (`_url`) into the JSON document
containing the requested data that is then returned to the auspice
client.

We want to move away from the JSON-manipulation and complicated logic
that follows on the auspice client for recognizing a new URL and
updating the browser's window state accordingly.

To start this process, we need to be able to access the `Response`
object returned from the server hosting the `/charon` web API.

Edit the `getDatasetFromCharon` function to return the `Response`
instead of the only the JSON object inside of the `Response`.

Because there is a higher level variable named `getDataset` that can
either contain the `getDatasetFromCharon` or the `getHardcodedData`
function, update `getHardcodedData` to return the `Response` instead of
the JSON object inside of the `Response`.

To prevent these changes from breaking the application, add a callback
to all usages of `getDatasetFromCharon`, `getHardcodedData`, and
`getDataset` which returns the JSON object from the `Response` data.