Update the window path using server response URL
Sometimes, partial URLs are sent to the `/charon/` web API endpoint.
The server sends back the correct data according to its best guess at
the complete URL.

Once the data are received from the server, update the client browser
window path to align with the HTTP endpoint the server followed to
complete the data request.

Note that this is a change of implementation of a currently existing
behavior. Previously, the server was injecting magic URLs into the
JSON object the client eventually received. We are no longer reading
URLs from the JSON object in this new implementation.