reduce the legend to only display valid values
The general idea is to ignore trait values on the tree which aren't valid, according to the `isValudeValid` function. We do not display such values in the legend, do not create colors for them, and do not create arcs for them on the map. Note that we do create a grey arc on each pie chart to represent the proportion of tips whose values are not "valid".