reinstate changing of URL via server redirect
Was accidentally removed during implementation of second tree URLs. Note that if auspice requests 2 datasets (as it interprets the URL as containing a second tree) and the server returns a redirect, then this is not interpreted. I can't see when this situation could arise, but it may want to be better addressed in the future.