Google analytics is opt-in
Starting from when auspice _was_ nextstrain.org, we hardcoded the google analytics (GA) key into auspice. This is no longer desirable. This commit changes behavior such that GA is only used if (i) the key is set as a build-time-extension (documentation to come) and (ii) auspice is run in production mode.