`auspice view` serves auspice from current directory if appropriate
We now serve a custom auspice build if one exists in the directory `auspice view` is run from. This streamlines generating custom auspice bundles and serving them locally. 

The option `--customBuild` is no longer used and will be removed in a future version. A warning is printed if it is specified. 

Documentation will be updated in a future commit.