simplify checking if node trait values are "valid"
Previously we stored the lookup of traits present in the tree for the current colorBy (e.g.) and their counts across the tree in a plain object, necessitating the use of the `UNDEFINED_VALUE` constant. As we now use a `Map` object, we can simply use `undefined` for this. This simplifies the logic around checking these values.