tree hover info displays supplied dates correctly
Previously all dates stated "Inferred Date" which was not correct. Augur currently creates confidence bounds for tips with known date values, but the confidence bounds are both the same. This commit returns to checking if the bounds are equal, and interpreting this as a defined date, not an inferred one.