interpret certain trait values on tips as known not inferred
Currently, the way augur exports confidence values for tips, it's largely impossible to know if a tip's trait which has 100% confidence is inferred or known (i.e. defined by the metadata). Since the majority of tips for which DTA is run have data, we assume that the value is provided.

This can be much improved upon resolution of https://github.com/nextstrain/augur/issues/386.