Shift buildUrls from getAvailable API to byline
Auspice (2.*) allows the `getAvailable` API request to define `buildUrl`s for datasets. These were displayed in the sidebar, above the choose-dataset dropdowns. 

This commit removes the sidebar display of this, as it's been super-seeded by the byline.

The commit also sets the `metadata.buildUrl` to be that defined in the `getAvailable` API response _if_ it has not been set via the dataset JSON, as the latter is to be considered the "canonical" way of defining this value.