Code cleanup
The "should i render" boolean approach is replaced with returning a react component when it's determined that something should be rendered, else returning `null` as a fallthrough at the end of the function (null is better than `<span/>`). 

The `renderLink` function is replaced by a component, which gets passed props rather than an object.

Looks like lots of changes but not that many really!