add `heroku-postbuild` script
This allows auspice to run on a heroku app by running the get-data and
get-narratives scripts on startup. This allows us to use heroku's
review-apps functionality. Note that `npm run build` is not automatically
run if `heroku-postbuild` is set. Note also that the HOST env variable
needs to be set to 0.0.0.0 on heroku for the auspice server to work.