Fix bug in computing state when nodes lack num_date attributes
The getMinCalDateViaTree function is used to initialize state of the date slider by getting the date of the root node. If the root node lacks the "num_date" attribute, the min date is set based on default date range. This default needs to be returned as a calendar date however.