move title (`<Helmet>`) higher up component tree
Tested on nextstrain.org locally resulted in two changes:

(a) moving the component higher up the tree as we want this to run for the Splash page, as well as for datasets. This allows nextstrain.org/groups/blab to show a title of `nextstrain / groups / blab`.

(b) connected it to redux state, as we modify the URL in certain situations (e.g. /flu/avian -> /flu/avian/h5n1/ha) and we want the head to change accordingly here. Note that this URL changes without the `<Main>` (dataset display) component rerunning, and is currently only in  nextstrain.org, not the default auspice server.)