improve temporal grid spacing on tree
Here we change the concept of minor grid lines for temporal trees to represent "sensible" periods of time rather than fitting a fixed number between major grid points. This will greatly aid in epidemiological understanding of trees. Major grid lines are also chosen using a new heuristic which attempts to pick sensible periods, and is responsive to screen sizes to avoid a squashed grid on small screens.

No modifications are made to the calculation of grids for divergence trees.