[bugfix] don't unnecessarily recompute colours on narrative page changes
A bug in `createStateFromQueryOrJSONs` caused the color scale to be recreated unnecessarily on each page change in a narrative. This was somewhat harmless, but meant that the map logic was confusing as it would tend to remove D3 elements & recreate them unnecessarily (as it detected a color-by change), which was different to the expected behavior outside of narratives.