[map] modify map zoom behavior
The previous implementation did not work on a lot of datasets (`demeIndices` and `demeData` seem to cause a lot of issues).

This commit breaks out the logic into a method `moveMapAccordingToData` which can be called from various places according to the react updates taking place. This is also the only place where the map is moved except for (a) the RESET ZOOM button and (b) when the user pans / zooms.

Effects:
* Loading a dataset with filters enabled zooms to the filtered data
* Narratives: changing filters between pages causes map zoom
* Narratives: changing geo res between pages causes map zoom

Also adds a number of console logs which should be cleaned up before release.