[map] automatically move map if user has not panned/zoomed
Already in narrative mode, if filters are enabled or geoRes changes, the map updates to fit the data in the screen. They weren't implemented for normal view, as moving the map can be annoying when you are interacting with the data.

Here we keep track of if a user has panned / zoomed the map at all (and this is reset upon pressing the "RESET ZOOM" button). If they are yet to interact with the map, then we do manipulate it when the filters (etc) change.

Note that clicking on the animation button "counts" as a user interaction. This avoids us constantly resizing the map during animation.