fix: ensure Lato font correctly falls back to sans-serif
Currently, some of the text elements, for example <a/>, are forced to be rendered using Lato font. When text contains Unicode characters (for example, from Chinese and Cyrillic alphabets) which are missing in our current version of Lato font, these characters are automatically substituted from other fonts.

However, this substitution is unpredictable (depends on browser implementation, platform, font and characters in question) and in some cases leads to inconsistent results.

This adds Helvetica and sans-serif alternatives to the places where only Lato was specified, thus at least ensuring that Unicode will properly falls back to sans font.