use query.label=<key>:<value> throughout code
This shifts the code usage to mirror the intended URL query. By removing all old references to `query.clade` we reduce potential confusion. By keeping a string type for `<key>:<value>` we simplify the code and keep it in sync with the mental model for how the URL query looks. These changes also allow zooming to root in narratives when no label (clade) is specified in the URL.