fix rect tree layout bug
The layout code defined both the stem & tee of a branch in a single SVG path, however the rendering code treated these differently and expected 2 paths. We were therefore rendering both paths via the code only intended for the stem.

This changes the display to draw stems over Ts. This should be improved by SVG gradients in future work.