Slightly revise entropy blending on branches
It had been such that a branch with 80% certainty in a particular location was completely gray. This felt off. In the revised behavior, we hit fully gray at about 50% certainty. This originally leaned quite heavily on gray because we were concerned that measures like 80% shouldn't actually be interpreted as "80%" due to over-confidence in the phylogeo reconstruction. Now, augur traits has --sampling-bias-correction, which corrects this at the source. Better to have auspice give an honest accounting at this point.