Emphasize branches on hover, including those with a gradient
Previously we changed the stroke colour of branches on hover. Here we extend this to handle linear gradients.

Previously, the on-hover stroke colour was somewhere between the "true" node colour and the "normal" stroke colour, and varied depending on the confidence attached to the colorBy at the node. This didn't look good with gradients, so I now emphasize the previously calculated stroke colour by increasing the saturation and making it slightly darker.