Improve which branches we display labels for
Auspice was build during a time when only "clade" and "aa" labels existed. We always showed labels for the former, and only showed the latter for branches with a large number of descendants.

Auspice now supports different labels, and this commit makes it so that any non "aa" label is always displayed. "aa" labels are shown if the branch has any other labels defined or if there are enough descendants.