Improve tree label font sizes
The font size should be a function of the total number of tips on the screen ("in-view"), regardless of their visibility. Renames variables to make the distinction clearer. There are still plenty of edge-cases where these don't look great, but this is a big improvement over the current version of auspice.