Add "end of narrative" slide to blocks.
Previously, there was a disconnect between the redux state's storage of narrative blocks, and the display of a "end of narrative" (EON) slide, which wasn't part of the blocks. PR1016 attempted to add a URL query to represent the EON, which proved problematic as we had a `blockIdx` outside of the range of the underlying blocks structure.

Here I take a different approach, by having auspice add the EON as a "normal" block on JSON parsing. Thus, we have a URL query etc without introducing lots of edge cases.