Place tip labels below tips
This changes the render order so that the SVG group holding tips is before (below) the tips, thus preventing labels being drawn on top of tips and preventing the on-hover behavior.

A better long-term solution for the problems with drawing tip labels on big trees may be found, but this will require different rendering behavior or different logic around which tip labels to draw.

Closes #1082.