Increase allowed bundlesizes
This is promped by a number of failing PRs which are slightly over the allowed bundlesizes
(these limits were arbitrarily picked to be very close to the bundlesizes at that time).
Analysis of the bundles (via `auspice build --analyzeBundle`) shows that there are numerous
modules and bits of code which could be in separate bundles -- the main bundle is especially
problematic. A separate issue can be made to bring these down.