Add opt-in genome download feature
When a corresponding FASTA file is placed in the same directory with auspice's
JSON file, for example a ncov_global.fasta is placed in the same directory with
ncov_global.json: the genome database handling functionality will kick in and
generate a file-based NeDB database at server startup time. A "Selected
Genomes (Fasta)" download button also then appears in side the "Download Data"
popup window. For dataset that doesn't have a corresponding FASTA file, this
functionality is inactive.