fix-loading-race-condition
Fixes a race-condition related to calculating the visible legend items when a new dataset was loading. In the order of operations, the colorScale is created before the tree visibility is calculated and therefore it's not possible to work out the legend items we wish to display at the point of color scale creation.