[frequencies] combine dispatch calls
Toggling "normalize frequencies" now results in only a single dispatch call. This reduces the chances of state becoming out-of-sync in future work, and will allow narratives to function better (if we ever use frequencies in narratives).

The "TOGGLE_X" actions work best if the react component will update appropriately when that state changes (or some middleware listens for it and updates redux state accordingly). As the frequencies are currently designed, however, the underlying matrix is updated independent of the frequencies panel, and the panel component is more of a plain-renderer.