Allow resampling from trait confidences
This commit is a proof-of-principle to resample color-by and/or geo-resolution values if available.

New values for discrete traits are picked for each node independently based on the discrete probabilities supplied by the dataset. For continuous values we take a uniform sample from the bounds provided by the dataset (often 95% CI).

This is achieved via the following key-board shortcuts:
"s" + "c": resample currently selected color-by
"s" + "r": resample currently selected geo-resolution.
Using capitol letters returns to the starting values rather than resampling.

There remain a number of "to-do"s before this can be merged.