Move to react-transition-group
"ReactTransitionGroup and ReactCSSTransitionGroup have been moved to the react-transition-group package that is maintained by the community. Its 1.x branch is completely API-compatible with the existing addons."

This commit moves Auspice to the 1.x branch of react-transition-group which doesn't require any API changes. The current version is v4, however given that we only use the library in a single place it's not high priority to update (the API has changed).