[bugfix] allow narratives to simultaneously zoom and change color
This fixes a bug where a narrative slide-change requests that the tree both zoom and change color.

This was introduced by 0639da85c4e1e1c343acbbe8d3231280ea8b68ca which, as part of the (now disabled) push for SVG gradient rendering of branches, moved the d3 call to update the stroke (branch color) into a separate code block, in order that it can behave differently for "T" and "S" parts of a branch. This code-block is not run when the tree zooms (that is the bug).

Here we partially revert that commit. This should be revisited when we re-enable SVG gradients.