Move to storing test narratives in the auspice repo
This commit signifies a shift to storing a set of narrative files in this repo, as opposed to obtaining the nextstrain.org narratives via a script.

The Nextstrain.org narratives are still obtainable (they are all stored in their own github repo) and the docs have been updated to show this.

By storing and maintaining "test narratives" in the auspice repo we can both provide examples of the capability of narratives and use them to test functionality and fix bugs. As the are stored in the `narratives` directory, they are available via Nextstrain Community URLs at nextstrain.org/community/narratives/nextstrain/auspice/...