Add a narratives snapshot test
This builds on our previous efforts to develop snapshot testing (not being run automatically at the moment). Here we dynamically compare snapshots for each page within a test-narrative. This code may be refactored so that we can easily test a wide range of narratives in this fashion.

Changes to the content of the narrative, including the addition of slides, should cause this test to fail, however the screenshots can be regenerated (see DEV_DOCS) without the need to modify the test script itself.