Improve redux state recalc on narrative slide changes. Closes #1191
This commit improves the creation of new state during narrative slide-changes which has a number of beneficial side-effects.

Previously we made a shallow-copy of the (redux) controls state when we changed slides, which resulted in props-comparison of nested state (i.e. objects within the controls state) being hard to reason with. This was problematic with the `temporalConfidence` object, which explains the approach employed in 0713e69b0707c92df806a575796e7205f08332ff to allow narrative slides to update CIs.

We now deep clone the controls state, allowing valid props comparison and allowing us to use this technique to decide when to render / remove CIs. (This is essentially a reversion of 0713e69b0707c92df806a575796e7205f08332ff, which was implemented as props comparisons were found to be unreliable).

This commit was motivated by, and fixes #1191. That bug was a result of the previous logic indicating that CIs should be removed on each branch-hover event. This meant that any subsequent tree-manipulation (such as clicking a branch to zoom) would be likely to be considered "in quick succession" to a previous manipulation and thus not employ an animation.

A test narrative is added.