Remove "experimental" terminology relating main-display-markdown
This commit represents a simplification of the page-change logic when navigating between pages (the `EXPERIMENTAL_showMainDisplayMarkdown` action is unnecessary as the `changePage` action will correctly set the panels).

Additionally we rename the react component to remove the "experimental" prefix as this has been in widespread use for a number of months now.