Improve logic around switching into & out of narrative mode
This commit represents a big simplification of how we toggle into and out of narrative mode. This happens in two places: the "exit narrative mode" button on the last narrative slide and the "explore the data" button (desktop only). The desired behavior is that we switch to "interactive mode" (i.e. with the controls sidebar) but with a "return to narrative" button (desktop only). The URL should change from a narrative pathname and n=<slide_number> query to the pathname of the dataset & query describing the view settings.

The code is simplified to use a consistent action for all the above behavior ("TOGGLE_NARRATIVE").  This allows multiple datasets to work appropriately, as well as window resizing triggering a switch to mobile display (or vice versa). 

This resolves a long-standing development paper-cut where the narrative URL would disappear when hot-reloading (due to the unmounting behavior). 


Co-authored-by: eharkins <eli.harkins@gmail.com>