Improve parsing of auspice URLs with a colon character
In preparation for new nextstrain.org functionality to allow
users to specify custom JSON locations as a URL within the
nextstrain.org URL (see https://github.com/nextstrain/nextstrain.org/pull/216)
we need to improve the logic auspice uses to identify dual trees.

dual trees are defined via <pathA>:<pathB> but we wish to allow a `:` charater
inside pathA (and/or pathB) and not interpret this as requesting two trees.