Update leaflet libraries
The leaflet library wasn't pinned correctly resulting in recent
installs of auspice using 1.7.1 which was incompatible with versions
of leaflet-gesture-handling allowed by auspice. See
https://github.com/nextstrain/nextstrain.org/issues/223 for more
information.

This commit fixes the problem by updating both libraries which allows us
to use an improved version of leaflet. The versions are now pinned
exactly preventing this issue in future.