Implement sidebar info pop-up boxes
This refactors the sidebar titles to allow an (optional) pop-up text box. This is useful to convey the functionality of certain features. The only current feature using it is the new filter-box.

Uses `react-tooltip` rather than implementing our own (as we did in the entropy panel for instance) which is more robust.

Currently disabled on mobile-view, which is left as a to-do item as it requires more testing to get right.