Allow strains to be searched as normal filters
With the introduction of a "filter" search box in the sidebar, we wanted to allow strain-search here which would allow us to get rid of the "strain search" box.

There are some slight differences with this new UI:

Positives: One can now select multiple strains, which is really cool. They are stored in the URL (comma separated list) and thus can be used in narratives.

Todo: The previous behavior of expanding the tip radii on partial search completion isn't part of this new UI, but it should be.