fix link to work when doc fetched by main rtd
this relative link did not work
after fetching this doc into
docs.nextstrain.org because the
folder structure is different from
this repo so this changes it to fit
with that of docs.nextstrain.org.
this means the relative link breaks
in this repo but this document is
not rendered in this repo's docs.
there are many more instances of this
which will need to recieve similar
edits and remain that way until
we have a unified toctree.