[bugfix] handle unknown values in color scales with provided scales
Colorscales have to deal with nodes with unknown values for the current colorScale (e.g. `getTraitFromNode(node, colorBy=undefined`). This was not being correctly handled in the case where we provided a scale in the dataset JSON. For certain cases (i.e. when traits had not been inferred for internal nodes) this resulted in internal nodes being coloured rather than grey as intended.