Refactor individual colorScales into functions
The `calcColorScale` function was overly long and consisted of a number of conditionals which decided on which color scale to create. This commit doesn't change the conditional logic at all, but it pulls the logic to create each scale into a separate function. This results in the `calcColorScale` function being readable.